@@ -1,6 +1,6 @@
 {
   "name": "gatsby-starter",
-  "version": "1.0.1",
+  "version": "1.2.2",
   "main": "index.js",
   "repository": "https://github.com/toastyboost/gatsby-starter",
   "author": "Dmitriy Sergeevich <dmitriy.dubotovkin@gmail.com>",
@@ -20,56 +20,47 @@
     "update": "ncu -u"
   },
   "dependencies": {
-    "click-outside-hook": "^1.1.0",
-    "gatsby": "^2.24.84",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "gatsby": "^3.1.1",
     "gatsby-plugin-alias-imports": "^1.0.5",
-    "gatsby-plugin-favicon": "^3.1.6",
-    "gatsby-plugin-google-analytics": "^2.3.19",
+    "gatsby-plugin-google-analytics": "^3.1.0",
     "gatsby-plugin-intl": "^0.3.3",
-    "gatsby-plugin-no-sourcemaps": "^2.2.1",
-    "gatsby-plugin-react-helmet": "^3.3.14",
+    "gatsby-plugin-manifest": "^3.1.0",
+    "gatsby-plugin-no-sourcemaps": "^3.1.0",
+    "gatsby-plugin-react-helmet": "^4.1.0",
     "gatsby-plugin-react-helmet-canonical-urls": "^1.4.0",
-    "gatsby-plugin-robots-txt": "^1.5.3",
-    "gatsby-plugin-sitemap": "^2.4.17",
-    "gatsby-plugin-styled-components": "^3.3.14",
-    "gatsby-plugin-typescript": "^2.4.24",
+    "gatsby-plugin-robots-txt": "^1.5.5",
+    "gatsby-plugin-sitemap": "^3.1.0",
+    "gatsby-plugin-styled-components": "^4.1.0",
+    "gatsby-plugin-typescript": "^3.1.0",
     "gatsby-plugin-typescript-checker": "^1.1.1",
-    "gatsby-plugin-webpack-size": "^1.0.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "gatsby-plugin-webfonts": "^2.0.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-helmet": "^6.1.0",
-    "react-redux": "^7.2.1",
-    "redux": "^4.0.5",
-    "redux-actions": "^2.6.5",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^5.2.0",
+    "styled-components": "^5.2.1",
     "styled-normalize": "^8.0.7"
   },
   "devDependencies": {
-    "@types/node": "^14.14.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
+    "@types/node": "^14.14.35",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.2",
     "@types/react-helmet": "^6.1.0",
-    "@types/react-redux": "^7.1.9",
-    "@types/redux": "^3.6.0",
-    "@types/redux-actions": "^2.6.1",
-    "@types/styled-components": "^5.1.4",
-    "@typescript-eslint/eslint-plugin": "^4.5.0",
-    "@typescript-eslint/parser": "^4.5.0",
+    "@types/styled-components": "^5.1.9",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
     "babel-eslint": "^10.1.0",
-    "babel-plugin-styled-components": "^1.11.1",
-    "babel-preset-gatsby": "^0.5.14",
-    "eslint": "^7.11.0",
-    "eslint-config-prettier": "^6.13.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-import-resolver-alias": "^1.1.2",
+    "babel-plugin-styled-components": "^1.12.0",
+    "babel-preset-gatsby": "^1.1.0",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-config-standard": "^16.0.2",
     "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.22.0",
     "eslint-utils": "^2.1.0",
-    "prettier": "^2.1.2",
-    "redux-logger": "^3.0.6",
-    "typescript": "^4.0.3"
+    "prettier": "^2.2.1",
+    "typescript": "^4.2.3"
   }
 }
