@@ -1,8 +1,7 @@
 import * as React from "react";
 
-import { Layout } from "~/ui/templates";
-
-import { Fonts, Async, Seo } from "~/ui/organisms";
+import { Layout } from "src/ui/templates";
+import { Seo } from "src/ui/organisms";
 
 const seoProps = {
   title: "Main",
@@ -13,10 +12,7 @@ const IndexPage: React.FC = () => {
   return (
     <>
       <Seo {...seoProps} />
-      <Layout>
-        <Async />
-        <Fonts />
-      </Layout>
+      <Layout>content</Layout>
     </>
   );
 };
