@@ -1,6 +1,5 @@
 import * as React from "react";
-
-import { FieldContainer, Field } from "./styles";
+import styled from "styled-components";
 
 type InputProps = {
   value: string;
@@ -9,8 +8,26 @@ type InputProps = {
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
 };
 
-export const Input: React.FC<InputProps> = ({ value, placeholder, isInvalid, onChange }) => (
+export const Input: React.FC<InputProps> = ({
+  value,
+  placeholder,
+  isInvalid,
+  onChange,
+}) => (
   <FieldContainer>
-    <Field value={value} placeholder={placeholder} aria-invalid={isInvalid} onChange={onChange} />
+    <Field
+      value={value}
+      placeholder={placeholder}
+      aria-invalid={isInvalid}
+      onChange={onChange}
+    />
   </FieldContainer>
 );
+
+const FieldContainer = styled.div``;
+
+const Field = styled.input`
+  &[aria-invalid="true"] {
+    border: 1px solid var(--warning);
+  }
+`;
