@@ -1,18 +1,23 @@
 import * as React from "react";
+import styled from "styled-components";
 
 import { Link } from "../Link";
-import { LogoContainer, LogoImg } from "./styles";
-
-import LogoSVG from "~/static/logo.svg";
 
 type LogoPrrops = {
   className?: string;
-}
+};
 
 export const Logo: React.FC<LogoPrrops> = ({ className }) => (
   <LogoContainer className={className}>
-    <Link to="/">
-      <LogoImg src={LogoSVG} />
-    </Link>
+    <Link to="/">Logo</Link>
   </LogoContainer>
 );
+
+const LogoContainer = styled.div`
+  width: 32px;
+  height: 32px;
+
+  &:hover {
+    opacity: 0.8;
+  }
+`;
